Moved scheduledTransfer ledger interaction to separate file
